Added aq add_aquilon_host, add_aurora_host, add_windows_host. Added aq add_dns_domain, show_dns_domain, del_dns_domain.
Default database changes that need to be added during migration
or afterwards by hand / manually:
- Archetype aurora
- Archetype windows
- Domain aurora_domain
- Domain windows_domain
- Vendor aurora_vendor
- Cpu aurora_cpu
- Model aurora_model with type aurora_node and zeroed MachineSpecs

Sneaking in to this commit:
- A fix to add_model to save the correct machine_type.
- A lock around del_host to serialize the command.  This was the
	only command that had obvious deadlock issues when
	being tested in parallel with other commands.
- Bumped all aqdb dependencies to the 0.4.7-1 in dist.
- Accessing specific location types (like location.rack) will
	return None if there is no relevant entry, instead of
	raising a KeyError.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 647972]

+++ 1.2.1/src/lib/python2.5/aquilon/server/commands/del_host.py
@@ -10,12 +10,18 @@
 """Contains the logic for `aq del host`."""
 
 
+from threading import Lock
+
 from twisted.python import log
 
+from aquilon.exceptions_ import ArgumentError, ProcessException
 from aquilon.server.broker import (format_results, add_transaction, az_check,
                                    BrokerCommand)
 from aquilon.server.dbwrappers.host import hostname_to_host
 from aquilon.server.processes import DSDBRunner
+
+
+delhost_lock = Lock()
 
 
 class CommandDelHost(BrokerCommand):
@@ -25,31 +31,45 @@ class CommandDelHost(BrokerCommand):
     @add_transaction
     @az_check
     def render(self, session, hostname, user, **arguments):
-        dbhost = hostname_to_host(session, hostname)
-        session.refresh(dbhost)
-        # Hack to make sure the machine object is refreshed in future queries.
-        dbmachine = dbhost.machine
-        session.refresh(dbmachine)
-        ip = None
-        for interface in dbmachine.interfaces:
-            if interface.boot:
-                ip = interface.ip
-        if not ip:
-            raise ArgumentError("No boot interface found for host to delete from dsdb.")
+        log.msg("Aquiring lock to attempt to delete %s" % hostname)
+        delhost_lock.acquire()
+        try:
+            log.msg("Aquired lock, attempting to delete %s" % hostname)
+            dbhost = hostname_to_host(session, hostname)
+            session.refresh(dbhost)
+            # Hack to make sure the machine object is refreshed in future queries.
+            archetype = dbhost.archetype.name
+            dbmachine = dbhost.machine
+            session.refresh(dbmachine)
+            ip = None
+            for interface in dbmachine.interfaces:
+                if interface.boot:
+                    ip = interface.ip
+            if not ip and archetype != 'aurora':
+                raise ArgumentError("No boot interface found for host to delete from dsdb.")
+    
+            for template in dbhost.templates:
+                log.msg("Before deleting host '%s', removing template '%s'"
+                        % (dbhost.fqdn, template.cfg_path))
+                session.delete(template)
+            session.delete(dbhost)
+            session.flush()
+    
+            if archetype != 'aurora':
+                try:
+                    dsdb_runner = DSDBRunner()
+                    dsdb_runner.delete_host_details(ip)
+                except ProcessException, e:
+                    raise ArgumentError("Could not remove host %s from dsdb: %s" %
+                            (hostname, e))
 
-        for template in dbhost.templates:
-            log.msg("Before deleting host '%s', removing template '%s'"
-                    % (dbhost.fqdn, template.cfg_path))
-            session.delete(template)
-        session.delete(dbhost)
-        session.flush()
+            # FIXME: Remove plenary template and profile.
 
-        dsdb_runner = DSDBRunner()
-        dsdb_runner.delete_host_details(ip)
+            session.refresh(dbmachine)
+        finally:
+            log.msg("Released lock from attempt to delete %s" % hostname)
+            delhost_lock.release()
 
-        # FIXME: Remove plenary template and profile.
-
-        session.refresh(dbmachine)
         return
